Add files via upload
Server can decrypt the received data. It can print it to screen and or to save it to file. 
User can create, fill, serialize, and deliver a dictionary to a server or send a text file to a server after creating it in Client. The pickling format of the dictionary can be binary, JSON, or XML.